# JasferAPI/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/emaaForlin/JasferInventoryServer/handlers
  Product:
    description: Product defines the structure for an API product
    properties:
      CreatedAt:
        format: date-time
        type: string
      Description:
        description: The description of the product
        maxLength: 1000
        type: string
      ID:
        description: The id for the product
        format: int64
        minimum: 1
        type: integer
      Name:
        description: The name of the product
        maxLength: 255
        type: string
      Price:
        description: The price of the product
        format: float
        minimum: 0.01
        type: number
      SKU:
        description: The intern unique code of the product
        type: string
      UpdatedAt:
        format: date-time
        type: string
    required:
    - Name
    - Price
    - SKU
    type: object
    x-go-package: github.com/emaaForlin/JasferInventoryServer/data
info:
  description: Documentation for Products API
  title: of Products API
  version: 1.0.0
paths:
  /products:
    get:
      description: Returns a list of products
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      tags:
      - products
    post:
      description: Create a new product
      operationId: createProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
        "501":
          $ref: '#/responses/errorResponse'
      tags:
      - products
  /products/{id}:
    delete:
      description: Deletes a product from the database
      operationId: deleteProduct
      parameters:
      - description: The ID of the product to delete from database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
      tags:
      - products
    put:
      description: 'Modifies a product '
      operationId: editProduct
      responses:
        "201":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  noContent:
    description: ""
  productsResponse:
    description: A list of products returns in the response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
swagger: "2.0"